+++ packages/core/src/devTools.ts
@@ -1,7 +1,6 @@
 import { Interpreter } from '.';
-import { IS_PRODUCTION } from './environment';
+import { AnyInterpreter } from './types';
 
-type AnyInterpreter = Interpreter<any, any, any>;
 type ServiceListener = (service: AnyInterpreter) => void;
 
 export interface XStateDevInterface {
@@ -40,7 +39,7 @@ function getDevTools(): XStateDevInterface | undefined {
 }
 
 export function registerService(service: AnyInterpreter) {
-  if (IS_PRODUCTION || !getGlobal()) {
+  if (!getGlobal()) {
     return;
   }