+++ packages/core/src/utils.ts
@@ -162,16 +162,11 @@ export function pathToStateValue(statePath: string[]): StateValue {
   return value;
 }
 
-export function mapValues<T, P>(
-  collection: { [key: string]: T },
-  iteratee: (
-    item: T,
-    key: string,
-    collection: { [key: string]: T },
-    i: number
-  ) => P
-): { [key: string]: P } {
-  const result: { [key: string]: P } = {};
+export function mapValues<T, P, O extends { [key: string]: T }>(
+  collection: O,
+  iteratee: (item: O[keyof O], key: keyof O, collection: O, i: number) => P
+): { [key in keyof O]: P } {
+  const result: Partial<{ [key in keyof O]: P }> = {};
 
   const collectionKeys = keys(collection);
   for (let i = 0; i < collectionKeys.length; i++) {
@@ -179,7 +174,7 @@ export function mapValues<T, P>(
     result[key] = iteratee(collection[key], key, collection, i);
   }
 
-  return result;
+  return result as { [key in keyof O]: P };
 }
 
 export function mapFilterValues<T, P>(
@@ -525,7 +520,8 @@ export function isObservable<T>(value: any): value is Subscribable<T> {
 }
 
 export const symbolObservable = (() =>
-  (typeof Symbol === 'function' && Symbol.observable) || '@@observable')();
+  (typeof Symbol === 'function' && (Symbol as any).observable) ||
+  '@@observable')();
 
 export function isMachine(value: any): value is StateMachine<any, any, any> {
   try {
@@ -646,7 +642,7 @@ export function reportUnhandledExceptionOnInvocation(
 }
 
 export function evaluateGuard<TContext, TEvent extends EventObject>(
-  machine: StateNode<TContext, any, TEvent>,
+  machine: StateNode<TContext, any, TEvent, any>,
   guard: Guard<TContext, TEvent>,
   context: TContext,
   _event: SCXML.Event<TEvent>,