+++ packages/xstate-vue/test/UseMachine.vue
@@ -11,10 +11,9 @@
 </template>
 
 <script lang="ts">
-import { PropType } from 'vue';
+import { defineComponent, PropType } from 'vue';
 import { useMachine } from '../src';
 import { Machine, assign, State } from 'xstate';
-import { watch } from 'vue';
 
 const context = {
   data: undefined
@@ -45,7 +44,7 @@ const fetchMachine = Machine<typeof context, any>({
   }
 });
 
-export default {
+export default defineComponent({
   props: {
     persistedState: {
       type: Object as PropType<State<any>>
@@ -63,5 +62,5 @@ export default {
     });
     return { state, send, service };
   }
-};
+});
 </script>